修正了RoomType主键类型；完成了Movie的新增和修改操作，但日期输入有bug还未修复；修改了部分页面外观；添加了若干注释。

+++ Models/RoomType.cs
@@ -4,7 +4,7 @@ namespace HotelBookingSystem.Models
 {
     public class RoomType
     {
-        public int Id { get; set; } // PK
+        public byte Id { get; set; } // PK
         
         [Required]
         [StringLength(255)]
